+++ app/src/main/java/com/zeni/core/data/local/database/dao/ItineraryDao.kt
@@ -1,11 +1,10 @@
-package com.zeni.core.data.database.dao
+package com.zeni.core.data.local.database.dao
 
-import android.icu.text.ListFormatter.Type.AND
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Query
 import androidx.room.Upsert
-import com.zeni.core.data.database.entities.ActivityEntity
+import com.zeni.core.data.local.database.entities.ActivityEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
