+++ app/src/main/java/com/zeni/core/data/mappers/Activity.kt
@@ -1,6 +1,6 @@
 package com.zeni.core.data.mappers
 
-import com.zeni.core.data.database.entities.ActivityEntity
+import com.zeni.core.data.local.database.entities.ActivityEntity
 import com.zeni.core.domain.model.Activity
 
 fun Activity.toEntity() = ActivityEntity(
