+++ app/src/main/java/com/zeni/core/data/mappers/TripImage.kt
@@ -1,7 +1,7 @@
 package com.zeni.core.data.mappers
 
 import androidx.core.net.toUri
-import com.zeni.core.data.database.entities.TripImageEntity
+import com.zeni.core.data.local.database.entities.TripImageEntity
 import com.zeni.core.domain.model.TripImage
 
 fun TripImage.toEntity() = TripImageEntity(
